feat(news-articles): refactoring the news article components to be reusable throughout the app
